+++ lib/private/Preview/Generator.php
@@ -20,7 +20,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-namespace OC;
+
+namespace OC\Preview;
 
 use OC\Files\View;
 use OCP\Files\File;
@@ -35,7 +36,7 @@ use OCP\Image;
 use OCP\IPreview;
 use OCP\Preview\IProvider;
 
-class Preview2 {
+class Generator {
 	//the thumbnail folder
 	const THUMBNAILS_FOLDER = 'thumbnails';
 
@@ -87,7 +88,7 @@ class Preview2 {
 	 * @return ISimpleFile
 	 * @throws NotFoundException
 	 */
-	public function getPreview($width = -1, $height = -1, $crop = false, $mode = Preview2::MODE_FILL) {
+	public function getPreview($width = -1, $height = -1, $crop = false, $mode = Generator::MODE_FILL) {
 		if (!$this->previewManager->isMimeSupported($this->file->getMimeType())) {
 			throw new NotFoundException();
 		}
